centralize gcloud utilities
