+++ fv3config/_datastore.py
@@ -7,6 +7,7 @@ import requests
 import appdirs
 import subprocess
 from ._exceptions import ConfigError, DataMissingError
+from gcloud import _is_gcloud_path
 
 if 'FV3CONFIG_CACHE_DIR' in os.environ:
     USER_CACHE_DIR = os.environ['FV3CONFIG_CACHE_DIR']
@@ -18,7 +19,6 @@ else:
 ARCHIVE_FILENAME = '2019-10-23-data-for-running-fv3gfs.tar.gz'
 ARCHIVE_FILENAME_ROOT = '2019-10-23-data-for-running-fv3gfs'
 ARCHIVE_URL = f'http://storage.googleapis.com/vcm-ml-public/{ARCHIVE_FILENAME}'
-GS_BUCKET_PREFIX = 'gs://'
 CACHE_PREFIX = 'fv3config-cache'
 
 FORCING_OPTIONS_DICT = {
@@ -101,53 +101,6 @@ def get_initial_conditions_directory(config):
     return resolve_option(config['initial_conditions'], INITIAL_CONDITIONS_OPTIONS_DICT)
 
 
-def link_or_copy_directory(source_path, target_path):
-    """Symbolically link or gsutil copy files in a source path to a target path"""
-    if is_gsbucket_url(source_path):
-        if gsutil_is_installed():
-            subprocess.check_call([
-                'gsutil', '-m', 'cp', '-r', os.path.join(source_path, '*'), target_path])
-        else:
-            logging.warning(
-                f'Optional dependency gsutil not found. Files in '
-                f'{source_path} will not be copied to {target_path}')
-    else:
-        link_directory(source_path, target_path)
-
-
-def link_directory(source_path, target_path):
-    """Recursively symbolic link the files in a source path into a target path.
-    """
-    for base_filename in os.listdir(source_path):
-        source_item = os.path.join(source_path, base_filename)
-        target_item = os.path.join(target_path, base_filename)
-        if os.path.isfile(source_item):
-            if os.path.exists(target_item):
-                os.remove(target_item)
-            os.symlink(source_item, target_item)
-        elif os.path.isdir(source_item):
-            if not os.path.isdir(target_item):
-                os.mkdir(target_item)
-            link_directory(source_item, target_item)
-
-
-def copy_file(source_path, target_path):
-    if is_gsbucket_url(source_path):
-        if gsutil_is_installed():
-            subprocess.check_call(['gsutil', 'cp', source_path, target_path])
-        else:
-            logging.warning(f'Optional dependency gsutil not found. File '
-                f'{source_path} will not be copied to {target_path}')
-    else:
-        shutil.copy(source_path, target_path)
-
-
-def gsutil_is_installed():
-    if shutil.which('gsutil') is None:
-        return False
-    else:
-        return True
-
 def check_if_data_is_downloaded():
     dirname = _get_internal_cache_dir()
     if not os.path.isdir(dirname) or len(os.listdir(dirname)) == 0:
@@ -194,10 +147,6 @@ def extract_data(archive_filename):
                 )
 
 
-def is_gsbucket_url(path):
-    return path.startswith(GS_BUCKET_PREFIX)
-
-
 def resolve_option(option, built_in_options_dict):
     """Determine whether a configuration dictionary option is a built-in option or
     not and return path to file or directory representing option. An option is
@@ -221,7 +170,7 @@ def resolve_option(option, built_in_options_dict):
             raise ConfigError(
                 f'The provided path {option} does not exist.'
             )
-    elif is_gsbucket_url(option):
+    elif _is_gcloud_path(option):
         return option
     else:
         if option in built_in_options_dict:
