+++ fv3config/gcloud.py
@@ -1,14 +1,25 @@
 import os
 import shutil
-import gcsfs
+from ._exceptions import DelayedImportError
+from fsspec.implementations.local import LocalFileSystem
+try:
+    import gcsfs
+except ImportError as err:
+    gcsfs = DelayedImportError(err)
+try:
+    import google.auth
+except ImportError as err:
+    google = DelayedImportError(err)
 
-GOOGLE_PROJECT = 'VCM-ML'
+LOCAL_FS = LocalFileSystem()
+GS_BUCKET_PREFIX = 'gs://'
+
 
 def _copy_directory(local_source_dir, dest_dir, fs=None):
     """Copy the contents of a local directory to a local or remote directory.
     """
     if fs is None:
-        fs = gcsfs.GCSFileSystem(project=GOOGLE_PROJECT)
+        fs = gcsfs.GCSFileSystem(project=_get_gcloud_project())
     for token in os.listdir(local_source_dir):
         source = os.path.join(os.path.abspath(local_source_dir), token)
         dest = os.path.join(os.path.abspath(dest_dir), token)
@@ -23,7 +34,7 @@ def _copy_directory(local_source_dir, dest_dir, fs=None):
 def _copy_file(source_filename, dest_filename, fs=None):
     """Copy a local or remote file to a local or remote target location."""
     if fs is None and (_is_gcloud_path(source_filename) or _is_gcloud_path(dest_filename)):
-        fs = gcsfs.GCSFileSystem(project=GOOGLE_PROJECT)
+        fs = gcsfs.GCSFileSystem(project=_get_gcloud_project())
     if not _is_gcloud_path(source_filename):
         if not _is_gcloud_path(dest_filename):
             shutil.copy2(source_filename, dest_filename, follow_symlinks=True)
@@ -37,4 +48,19 @@ def _copy_file(source_filename, dest_filename, fs=None):
 
 
 def _is_gcloud_path(path):
-    return path[:5] == 'gs://'
+    return path.startswith(GS_BUCKET_PREFIX)
+
+
+def _get_gcloud_project():
+    return os.environ.get(google.auth.environment_vars.PROJECT)
+
+
+def _get_fs(path):
+    if _is_gcloud_path(path):
+        return _get_gcloud_fs()
+    else:
+        return LOCAL_FS
+
+
+def _get_gcloud_fs():
+    return gcsfs.GCSFileSystem(project=_get_gcloud_project())
